+++ earth_wallpaper/interfaces/utils/settings.py
@@ -3,6 +3,13 @@ from PySide6.QtWidgets import QApplication
 import os
 
 
+def trans(i: str) -> bool:
+    if i == "true" or i == "True":
+        return True
+    if i == "false" or i == "False":
+        return False
+
+
 class Settings(object):
     def __init__(self):
         config_dir = QStandardPaths.writableLocation(QStandardPaths.ConfigLocation)
@@ -10,7 +17,7 @@ class Settings(object):
         self.settings = QSettings(config_path, QSettings.IniFormat)
 
     def proxies(self) -> dict:
-        type_list = ["None", "http", "socks"]
+        type_list = ["None", "http", "socks5h"]
         prx_type = type_list[int(self.settings.value("System/proxy"))]
         prx_add = self.settings.value("System/proxyAdd")
         prx_port = self.settings.value("System/proxyPort")
@@ -29,6 +36,36 @@ class Settings(object):
 
     def earth_size(self) -> int:
         return int(self.settings.value("APP/earthSize"))
+
+    def apikey(self) -> str:
+        return self.settings.value("APP/apikey")
+
+    def general(self):
+        return trans(str(self.settings.value("APP/General")))
+
+    def anime(self):
+        return trans(str(self.settings.value("APP/Anime")))
+
+    def people(self):
+        return trans(str(self.settings.value("APP/People")))
+
+    def sfw(self):
+        return trans(str(self.settings.value("APP/SFW")))
+
+    def sketchy(self):
+        return trans(str(self.settings.value("APP/Sketchy")))
+
+    def nsfw(self):
+        return trans(str(self.settings.value("APP/NSFW")))
+
+    def sorting(self):
+        return self.settings.value("APP/sorting").lower().replace(" ", "_")
+
+    def searchkey(self):
+        return self.settings.value("APP/searchkey")
+
+    def color(self):
+        return self.settings.value("APP/color")
 
     @staticmethod
     def desktop_res() -> (int, int):