+++ earth_wallpaper/interfaces/utils/wallpaper24Common.py
@@ -1,9 +1,14 @@
 from .sunCalculator import SunCalculator, DateTime
 from .platformInfo import PlatformInfo
+from geopy.geocoders import Nominatim
 import os
 import json
 import time
 import requests
+import logging
+
+
+logger = logging.getLogger(__name__)
 
 
 def find_first_json(dir_):
@@ -11,6 +16,7 @@ def find_first_json(dir_):
     for file in files:
         if file.endswith(".json"):
             return file
+    logger.fatal("找不到json文件")
 
 
 class Wallpaper24Common(object):
@@ -37,38 +43,51 @@ class Wallpaper24Common(object):
         i = 0
         while i < 3:
             try:
-                ip = session.get("https://checkip.amazonaws.com/",
-                                 timeout=5).text.strip()
-                print(ip)
-                loc = session.get("https://ipapi.co/{}/json/".format(ip),
-                                  timeout=5).json()
-                print(loc)
-                latitude = float(loc["latitude"])
-                longitude = float(loc["longitude"])
+                # ip = session.get("https://checkip.amazonaws.com/",
+                #                  timeout=5).text.strip()
+                # print(ip)
+                # loc = session.get("https://ipapi.co/{}/json/".format(ip),
+                #                   timeout=5).json()
+
+                # loc = {"latitude": "29.5689", "longitude": "106.5577"}
+                #
+                # print(loc)
+                # latitude = float(loc["latitude"])
+                # longitude = float(loc["longitude"])
+
+                geolocator = Nominatim(user_agent='Mozilla/5.0 (Windows NT 10.0; Win64; x64) AppleWebKit/537.36 '
+                                                  '(KHTML, like Gecko) Chrome/75.0.3770.90 Safari/537.36 '
+                                                  '@earth-wallpaper-ext')
+                location = geolocator.geocode("重庆市")
+                logger.info(location.address)
+                logger.info(location.latitude, location.longitude)
+                latitude = float(location.latitude)
+                longitude = float(location.longitude)
+
                 i = 3
                 return self.calculate_sun(latitude, longitude)
             except ConnectionResetError:
-                print(f"本机IP获取失败，第{i + 1}次重试")
+                logger.warning(f"本机IP获取失败，第{i + 1}次重试")
                 if i == 3:
                     return self.calculate_time(5, 18)
                 else:
                     i += 1
             except KeyError:
-                print("API响应错误，使用默认时间")
+                logger.warning("API响应错误，使用默认时间")
                 i = 3
                 return self.calculate_time(5, 18)
             except TypeError:
-                print("该IP获取不到地理坐标，使用默认时间")
+                logger.warning("该IP获取不到地理坐标，使用默认时间")
                 i = 3
                 return self.calculate_time(5, 18)
             except requests.exceptions.ReadTimeout:
-                print(f"请求超时,第{i + 1}次重试...")
+                logger.warning(f"请求超时,第{i + 1}次重试...")
                 if i == 3:
                     return self.calculate_time(5, 18)
                 else:
                     i += 1
             except requests.exceptions.ConnectionError:
-                print("无网络连接,使用默认时间")
+                logger.warning("无网络连接,使用默认时间")
                 i = 3
                 return self.calculate_time(5, 18)
 
@@ -96,7 +115,7 @@ class Wallpaper24Common(object):
         hour = time.localtime(time.time()).tm_hour
 
         if 'sunriseImageList' not in theme and hour in sunrise:
-            print('sunrise day')
+            logger.info('sunrise day')
             newDay = day + sunrise
             num = len(newDay) // len(theme["dayImageList"])
             index = newDay.index(hour) // num
@@ -108,7 +127,7 @@ class Wallpaper24Common(object):
                 return data
 
         if 'sunsetImageList' not in theme and hour in sunset:
-            print('sunset night')
+            logger.info('sunset night')
             newNight = night + sunset
             num = len(newNight) // len(theme["nightImageList"])
             index = newNight.index(hour) // num
@@ -120,7 +139,7 @@ class Wallpaper24Common(object):
                 return data
 
         if hour in sunrise:
-            print('sunrise')
+            logger.info('sunrise')
             num = len(sunrise) // len(theme["sunriseImageList"])
             index = sunrise.index(hour) // num
             if index >= len(theme["sunriseImageList"]):
@@ -130,7 +149,7 @@ class Wallpaper24Common(object):
                 data = fp.read()
                 return data
         elif hour in day:
-            print('day')
+            logger.info('day')
             num = len(day) // len(theme["dayImageList"])
             index = day.index(hour) // num
             if index >= len(theme["dayImageList"]):
@@ -140,7 +159,7 @@ class Wallpaper24Common(object):
                 data = fp.read()
             return data
         elif hour in sunset:
-            print('sunset')
+            logger.info('sunset')
             num = len(sunset) // len(theme["sunsetImageList"])
             index = sunset.index(hour) // num
             if index >= len(theme["sunsetImageList"]):
@@ -150,7 +169,7 @@ class Wallpaper24Common(object):
                 data = fp.read()
             return data
         elif hour in night:
-            print('night')
+            logger.info('night')
             num = len(night) // len(theme["nightImageList"])
             index = night.index(hour) // num
             if index >= len(theme["nightImageList"]):
@@ -160,7 +179,7 @@ class Wallpaper24Common(object):
                 data = fp.read()
             return data
         else:
-            print("Error")
+            logger.error("Error")
 
     def run(self):
         self.check()