+++ earth_wallpaper/utils/setWallpaper.py
@@ -1,10 +1,21 @@
 from .platformInfo import PlatformInfo
+from PySide6.QtCore import QSettings, QStandardPaths
+import logging
 import os
+
+logger = logging.getLogger(__name__)
 
 
 def set_wallpaper(file):
+    config_dir = QStandardPaths.writableLocation(QStandardPaths.ConfigLocation)
+    config_path = os.path.join(config_dir, "earth-wallpaper/earth-wallpaper.conf")
+    settings = QSettings(config_path, QSettings.IniFormat)
+    scripts = settings.value("System/scripts")
+    if not len(scripts) == 0:
+        os.system(scripts + " " + file)
+        return
     sys = PlatformInfo().get_os()
-
+    logger.info(f"当前系统为{sys}")
     if sys == "WINDOWS":
         import win32api, win32gui, win32con
         key = win32api.RegOpenKeyEx(win32con.HKEY_CURRENT_USER,
@@ -18,6 +29,7 @@ def set_wallpaper(file):
 
     elif sys == "LINUX":
         de = os.getenv('XDG_CURRENT_DESKTOP')
+        logger.info(f"当前桌面环境为{de}")
         if de == "Deepin":
             # import dbus
             from .deepinWallPaperHandler import exec_setting
@@ -75,4 +87,4 @@ def set_wallpaper(file):
         elif de == 'LXDE':
             os.system("pcmanfm -w {}".format(file))
         else:
-            print("该桌面环境暂不支持")
+            logger.info("当前桌面环境暂不支持")