+++ setup.sh
@@ -1,0 +1,5 @@
+#!/bin/bash
+
+export PYTHONPATH=/home/sky198/.local/lib/python3.7/site-packages
+
+python3 setup.py install --prefix=/home/sky198/.local --record install.log